fix legend on lat/long maps: OR, incid by seas&wk
for all three maps:
changed legend colors to "bin" variable because the "legend" variable was misrepresenting the OR/attack rate bin factors
removed decreasing=TRUE in label sorting because it caused the labels to applied incorrectly
reversed the factor levels so blue would refer to lower OR and incidence values and red would be for high values
removed the "legend" variable from the dataframes because it was no longer needed for plotting the color scale
fixed labels dataset to "mergethree" because the other dataset would allow missing values to appear
show all OR/incidence factors, will not drop unused factors, which will keep the legends the same regardless of the presence of a certain factor on the map

To do:
take log(OR)

Notes:
The incidence attack rates seem unusually low when the week and season ILI & popstat values are summed and percentages are calculated. Look at the original mysql data to see if attack rates fall below 1% as they do here. This needs to be fixed