clean up master branch
setting up files that will be pushed to new github account, but not quite sure how I will push only certain files to the remote.
perhaps have a WIP branch with all of the files and the clean github files on the master branch?

minor changes to descriptive_age_dynamics.lyx outline