explore increased SD for 'region' method of retrozOR
F4_zOR_region.py
F4_zOR_region_reglvl.py
- added some code to print dict values but removed
it for this commit
- it seems that the 'region' incidence peak based
retrozOR method has greater SDs across all regions
than the 'nation' incidence peak based method
because the metric is capturing the epidemic growth
phase of the season. this part of the season is likely
to have more monotonic changes in zOR than if the
metric captured part of the epidemic growth phase
and the epidemic peak. perhaps this monotonicity
in zOR is what is causing the values of the mean
retrozOR to have greater standard deviation across
the board