ILINet fxns - non-nested with data as sole args
functions.py
- changed structure of functions for ILINet analyses
(OR, zOR, week plotting dicts, classif zOR) so that
only the OR function reads in the data files. The
other subsequent functions build off of the data
objects created in that initial function. The functions
must be run sequentially in the main script in order
to generate the data objects in the correct order (to
feed into the next function called), but they are no
longer nested in the funcitons.py script.
- Think about removing the ILINet and SDI data distinctions
in functions; instead, have an area in functions.py
where the used parameters are set (1. dummy variable
name for plotting seasons list in all scripts 2. keep
ILINet and SDI plotting seasons lists as separate names
3. add dummy variable to section where used parameters
are set and call the ILINet or SDI plotting seasons lists
as necessary)