export data -> new fxn update
export_zOR_classif.py
- updated to reflect new function dependencies in
functions.py
- as written, SDI or ILINet classification values
may only be exported when the other data import
and export sections are commented out. This is
because the pseasons variable in functions.py
must be assigned solely to the SDI data or ILINet
data season numbers.

functions.py
- updated state-related functions, which had
errors because they had been previously
updated for the new function dependences, but
not tested
- removed incorrect comments