state/reg incid retro pkwk based classif
previously, most analyses had been performed at the
"nation level retrospective peak week based classification",
which means that the peak incidence week of the nation level
incidence curve was used to define the retrospective zOR
classification period for the state and regional levels. In
an attempt to make the analyses more compact, we decided
to explore the option of state/reg incidence peaks defining
state/reg retrospective classifications. In this new
framework, the retrospective period is different for every
state/region within a season.

Results do not appear to be the same at first glance.